requirements.txt
python-dotenv==1.1.1
youtube-transcript-api==1.2.2
langchain==0.3.27
langchain-chroma==0.2.5
langchain-core
fastapi[standard]
uvicorn
pydantic
langchain-google-genai